index: refactor functions
