+++ index/lib.h
@@ -28,6 +28,7 @@
  * | File                 | Description                 |
  * | :------------------- | :-------------------------- |
  * | index/config.c       | @subpage index_config       |
+ * | index/functions.c    | @subpage index_functions    |
  * | index/index.c        | @subpage index_index        |
  * | index/observer.c     | @subpage index_observer     |
  * | index/private_data.c | @subpage index_private_data |
@@ -40,6 +41,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
+#include "mx.h"
 #include "core/lib.h"
 #include "shared_data.h" // IWYU pragma: keep
 
@@ -66,5 +68,29 @@ void mutt_set_header_color(struct Mailbox *m, struct Email *e);
 void mutt_update_index(struct Menu *menu, struct Context *ctx, enum MxStatus check, int oldcount, struct IndexSharedData *shared);
 struct MuttWindow *index_pager_init(void);
 void index_pager_shutdown(struct MuttWindow *dlg);
+int mutt_dlgindex_observer(struct NotifyCallback *nc);
+bool check_acl(struct Mailbox *m, AclFlags acl, const char *msg);
+int ci_next_undeleted(struct Mailbox *m, int msgno);
+void update_index(struct Menu *menu, struct Context *ctx, enum MxStatus check, int oldcount, const struct IndexSharedData *shared);
+void change_folder_mailbox(struct Menu *menu, struct Mailbox *m, int *oldcount, struct IndexSharedData *shared, bool read_only);
+void collapse_all(struct Context *ctx, struct Menu *menu, int toggle);
+void change_folder_string(struct Menu *menu, char *buf, size_t buflen, int *oldcount, struct IndexSharedData *shared, bool *pager_return, bool read_only);
+int ci_previous_undeleted(struct Mailbox *m, int msgno);
+int ci_first_message(struct Mailbox *m);
+void resort_index(struct Context *ctx, struct Menu *menu);
+int mx_toggle_write(struct Mailbox *m);
+extern const struct Mapping IndexNewsHelp[];
+struct Mailbox *change_folder_notmuch(struct Menu *menu, char *buf, int buflen, int *oldcount, struct IndexSharedData *shared, bool read_only);
+
+/**
+ * typedef CheckFlags - Checks to perform before running a function
+ */
+typedef uint8_t CheckFlags;       ///< Flags, e.g. #CHECK_IN_MAILBOX
+#define CHECK_NO_FLAGS         0  ///< No flags are set
+#define CHECK_IN_MAILBOX (1 << 0) ///< Is there a mailbox open?
+#define CHECK_MSGCOUNT   (1 << 1) ///< Are there any messages?
+#define CHECK_VISIBLE    (1 << 2) ///< Is the selected message visible in the index?
+#define CHECK_READONLY   (1 << 3) ///< Is the mailbox readonly?
+#define CHECK_ATTACH     (1 << 4) ///< Is the user in message-attach mode?
 
 #endif /* MUTT_INDEX_LIB_H */
